feat: Add GraphQL fragments and modularize project and issue retrieval
- Introduced common GraphQL fragments for issues, pull requests, and project items in `fragments.js` to streamline data fetching.
- Created separate modules for handling project and issue retrieval in `projects.js` and `issues.js`, improving code organization and maintainability.
- Updated `index.js` to import and utilize the new retrieval functions, enhancing clarity and separation of concerns.
- Modified the GitHub Actions workflow to trigger on feature branches, improving CI/CD flexibility.
