+++ .github/workflows/main.yml
@@ -2,6 +2,7 @@ on:
   push:
     branches:
       - main
+      - 'feat-*'
   workflow_dispatch:
 
 jobs:
